MF-334 - Move policies DB methods in a separate file
