+++ auth/api/grpc/endpoint_test.go
@@ -45,7 +45,7 @@ func newService() auth.Service {
 	idProvider := uuid.NewMock()
 	t := jwt.New(secret)
 
-	return auth.New(nil, nil, nil, repo, nil, idProvider, t, loginDuration)
+	return auth.New(nil, nil, nil, repo, nil, nil, idProvider, t, loginDuration)
 }
 
 func startGRPCServer(svc auth.Service, port int) {
