+++ auth/orgs.go
@@ -83,33 +83,6 @@ type OrgGroupsPage struct {
 	OrgGroups []OrgGroup
 }
 
-type GroupsPolicy struct {
-	GroupID  string
-	MemberID string
-	Policy   string
-}
-
-type GroupInvitationByID struct {
-	MemberID string
-	Policy   string
-}
-
-type GroupInvitationByEmail struct {
-	Email  string
-	Policy string
-}
-
-type GroupMember struct {
-	MemberID string
-	Email    string
-	Policy   string
-}
-
-type GroupMembersPage struct {
-	PageMetadata
-	GroupMembers []GroupMember
-}
-
 type OrgMember struct {
 	MemberID  string
 	OrgID     string
@@ -177,18 +150,6 @@ type Orgs interface {
 	// ListOrgGroups retrieves groups assigned to an org identified by orgID.
 	ListOrgGroups(ctx context.Context, token, orgID string, pm PageMetadata) (GroupsPage, error)
 
-	// CreateGroupMembers creates group members.
-	CreateGroupMembers(ctx context.Context, token, groupID string, giByEmails ...GroupInvitationByEmail) error
-
-	// ListGroupMembers retrieves page of group members.
-	ListGroupMembers(ctx context.Context, token, groupID string, pm PageMetadata) (GroupMembersPage, error)
-
-	// UpdateGroupMembers updates group members.
-	UpdateGroupMembers(ctx context.Context, token, groupID string, giByEmails ...GroupInvitationByEmail) error
-
-	// RemoveGroupMembers removes group members.
-	RemoveGroupMembers(ctx context.Context, token, groupID string, memberIDs ...string) error
-
 	// Backup retrieves all orgs, org members and org groups. Only accessible by admin.
 	Backup(ctx context.Context, token string) (Backup, error)
 
@@ -254,19 +215,4 @@ type OrgRepository interface {
 
 	// RetrieveAllOrgGroups retrieves all org groups.
 	RetrieveAllOrgGroups(ctx context.Context) ([]OrgGroup, error)
-
-	// SaveGroupMembers saves group members.
-	SaveGroupMembers(ctx context.Context, groupID string, giByIDs ...GroupInvitationByID) error
-
-	// RetrieveGroupMember retrieves group policy for a user.
-	RetrieveGroupMember(ctc context.Context, gp GroupsPolicy) (string, error)
-
-	// RetrieveGroupMembers retrieves page of group members.
-	RetrieveGroupMembers(ctx context.Context, groupID string, pm PageMetadata) (GroupMembersPage, error)
-
-	// RemoveGroupMembers removes group members.
-	RemoveGroupMembers(ctx context.Context, groupID string, memberIDs ...string) error
-
-	// UpdateGroupMembers updates group members.
-	UpdateGroupMembers(ctx context.Context, groupID string, giByIDs ...GroupInvitationByID) error
 }
